Add --cli-hide-cmd option
Some commands in obmc-yadro-cli require the command name
to be hidden in the help (e.g. for `bmc ntpconfig help`),
while others require the name to be shown in the help
(e.g. `bmc ifconfig help ntp`).

This commit extends the `--cli` option with another
version, `--cli-hide-cmd`. The former one will only
hide the banner while the latter will hide both the
banner and the command name in the command help.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>